requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.4
django-crispy-forms==2.0
django-environ==0.10.0
django-phonenumber-field==7.1.0
django-recaptcha==3.0.0
django-tinymce==3.6.1
fontawesomefree==6.4.2
phonenumbers==8.13.18
sqlparse==0.4.4
tzdata==2023.3
